+++ src/App.js
@@ -1,15 +1,11 @@
 import React, {useState} from 'react';
 import './App.css';
 import List              from './components/List';
+import Form              from './components/Form';
 
 export default function App() {
   const [todoData, setTodoData] = useState([]);
   const [value, setValue] = useState('');
-
-  const handleChange = (e) => {
-    console.log('e', e.target.value);
-    setValue(e.target.value);
-  };
 
   const handleSubmit = (e) => {
     e.preventDefault();
@@ -35,18 +31,7 @@ export default function App() {
 
         <List todoData={todoData} setTodoData={setTodoData}/>
 
-        <form style={{display: 'flex'}} onSubmit={handleSubmit}>
-          <input type="text"
-                 name="value"
-                 style={{flex: '10', padding: '5px'}}
-                 placeholder="해야 할 일을 입력하세요."
-                 value={value}
-                 onChange={handleChange}/>
-          <input type="submit"
-                 value="입력"
-                 className="btn"
-                 style={{flex: '1'}}/>
-        </form>
+        <Form handleSubmit={handleSubmit} value={value} setValue={setValue}/>
       </div>
     </div>
   );
